Feature loading issue rectified

+++ lib/feature_computation/load_features.py
@@ -71,9 +71,9 @@ class LoadFeatures:
                 fv_ = fv_.T
             feature_vectors_[speaker_id_][split_id_] = np.array(fv_, ndmin=2)
             if np.size(X_combined_)<=1:
-                X_combined_ = np.array(fv_, dtype=np.float32)
+                X_combined_ = np.array(fv_)
             else:
-                X_combined_ = np.append(X_combined_, np.array(fv_, dtype=np.float32), axis=0)
+                X_combined_ = np.append(X_combined_, np.array(fv_), axis=0)
 
             split_count_ += 1
             print(f'Loading features ({split_count_}/{len(self.INFO.keys())})', end='\r', flush=True)
